bug fix: 4queenが動作しなかった問題を修正。(mem_encode.cで、デコード時の整数の読み込み部分を修正) bug fix: 整数データアトムをCのint型で扱っていた部分をlongに修正

+++ slim/src/dumper.c
@@ -236,7 +236,7 @@ static BOOL dump_data_atom(LmnPort port,
   switch (attr) {
   case  LMN_INT_ATTR:
     {
-      char *s = int_to_str((int)data);
+      char *s = int_to_str((long)data);
       port_put_raw_s(port, s);
       LMN_FREE(s);
     }
